requirements.txt

# Web 
beautifulsoup4
googlesearch-python
requests

ollama